Add consistency to handling of column default. Fixes #315

+++ src/main/java/org/datanucleus/store/rdbms/mapping/column/BigIntColumnMapping.java
@@ -193,7 +193,7 @@ public class BigIntColumnMapping extends AbstractColumnMapping
             {
                 if (value instanceof Character)
                 {
-                    ps.setInt(param, (Character)value);
+                    ps.setLong(param, (Character)value);
                 }
                 else if (value instanceof String)
                 {
